+++ pkg/controller/scylladbdatacenter/resource.go
@@ -65,6 +65,9 @@ const (
 	alternatorInsecurePortName = "alternator"
 	alternatorTLSPort          = 8043
 	alternatorTLSPortName      = "alternator-tls"
+
+	nodeExporterPort     = 9100
+	nodeExporterPortName = "node-exporter"
 )
 
 var (
@@ -283,8 +286,8 @@ func getServicePorts(sdc *scyllav1alpha1.ScyllaDBDatacenter) ([]corev1.ServicePo
 			Port: scylla.DefaultScyllaDBManagerAgentMetricsPort,
 		},
 		{
-			Name: "node-exporter",
-			Port: 9100,
+			Name: nodeExporterPortName,
+			Port: nodeExporterPort,
 		},
 		{
 			Name: portNameThrift,
@@ -334,7 +337,7 @@ func getServicePorts(sdc *scyllav1alpha1.ScyllaDBDatacenter) ([]corev1.ServicePo
 
 // StatefulSetForRack make a StatefulSet for the rack.
 // existingSts may be nil if it doesn't exist yet.
-func StatefulSetForRack(rack scyllav1alpha1.RackSpec, sdc *scyllav1alpha1.ScyllaDBDatacenter, existingSts *appsv1.StatefulSet, sidecarImage string, rackOrdinal int, inputsHash string) (*appsv1.StatefulSet, error) {
+func StatefulSetForRack(rack scyllav1alpha1.RackSpec, sdc *scyllav1alpha1.ScyllaDBDatacenter, existingSts *appsv1.StatefulSet, sidecarImage string, nodeExporterImage string, rackOrdinal int, inputsHash string) (*appsv1.StatefulSet, error) {
 	selectorLabels, err := naming.RackSelectorLabels(rack, sdc)
 	if err != nil {
 		return nil, fmt.Errorf("can't get selector labels: %w", err)
@@ -344,6 +347,7 @@ func StatefulSetForRack(rack scyllav1alpha1.RackSpec, sdc *scyllav1alpha1.Scylla
 	if err != nil {
 		return nil, fmt.Errorf("can't get version of image %q: %w", sdc.Spec.ScyllaDB.Image, err)
 	}
+	scyllaDBSemver := semver.NewScyllaVersion(scyllaDBVersion)
 
 	if sdc.Spec.RackTemplate != nil {
 		rack = applyRackTemplateOnRackSpec(sdc.Spec.RackTemplate, rack)
@@ -469,7 +473,7 @@ func StatefulSetForRack(rack scyllav1alpha1.RackSpec, sdc *scyllav1alpha1.Scylla
 		minReadySeconds = int(*sdc.Spec.MinReadySeconds)
 	}
 
-	scyllaContainerPorts, err := containerPorts(sdc)
+	scyllaContainerPorts, err := containerPorts(sdc, scyllaDBSemver)
 	if err != nil {
 		return nil, fmt.Errorf("can't get scylla container ports: %w", err)
 	}
@@ -1109,10 +1113,14 @@ wait
 		sts.Spec.Template.Spec.Containers = append(sts.Spec.Template.Spec.Containers, *agentContainer)
 	}
 
+	if nodeExporterContainer := getScyllaDBNodeExporterContainer(nodeExporterImage, scyllaDBSemver); nodeExporterContainer != nil {
+		sts.Spec.Template.Spec.Containers = append(sts.Spec.Template.Spec.Containers, *nodeExporterContainer)
+	}
+
 	return sts, nil
 }
 
-func containerPorts(sdc *scyllav1alpha1.ScyllaDBDatacenter) ([]corev1.ContainerPort, error) {
+func containerPorts(sdc *scyllav1alpha1.ScyllaDBDatacenter, sv semver.ScyllaVersion) ([]corev1.ContainerPort, error) {
 	ports := []corev1.ContainerPort{
 		{
 			Name:          "intra-node",
@@ -1139,13 +1147,16 @@ func containerPorts(sdc *scyllav1alpha1.ScyllaDBDatacenter) ([]corev1.ContainerP
 			ContainerPort: scylla.DefaultScyllaDBMetricsPort,
 		},
 		{
-			Name:          "node-exporter",
-			ContainerPort: 9100,
-		},
-		{
 			Name:          "thrift",
 			ContainerPort: 9160,
 		},
+	}
+
+	if !sv.SupportFeatureSafe(semver.ScyllaDBVersionWithoutNodeExporter) {
+		ports = append(ports, corev1.ContainerPort{
+			Name:          nodeExporterPortName,
+			ContainerPort: nodeExporterPort,
+		})
 	}
 
 	if sdc.Spec.ScyllaDB.AlternatorOptions != nil {
@@ -1459,6 +1470,48 @@ exec scylla-manager-agent \
 	}
 
 	return cnt, nil
+}
+
+func getScyllaDBNodeExporterContainer(image string, sv semver.ScyllaVersion) *corev1.Container {
+	if !sv.SupportFeatureSafe(semver.ScyllaDBVersionWithoutNodeExporter) {
+		return nil
+	}
+
+	return &corev1.Container{
+		Name:            naming.ScyllaDBNodeExporterContainerName,
+		Image:           image,
+		ImagePullPolicy: corev1.PullIfNotPresent,
+		Ports: []corev1.ContainerPort{
+			{
+				Name:          nodeExporterPortName,
+				ContainerPort: nodeExporterPort,
+			},
+		},
+		LivenessProbe: &corev1.Probe{
+			ProbeHandler: corev1.ProbeHandler{
+				TCPSocket: &corev1.TCPSocketAction{
+					Port: apimachineryutilintstr.FromInt32(nodeExporterPort),
+				},
+			},
+		},
+		Resources: corev1.ResourceRequirements{
+			Limits: corev1.ResourceList{
+				corev1.ResourceCPU:    resource.MustParse("10m"),
+				corev1.ResourceMemory: resource.MustParse("50Mi"),
+			},
+			Requests: corev1.ResourceList{
+				corev1.ResourceCPU:    resource.MustParse("10m"),
+				corev1.ResourceMemory: resource.MustParse("50Mi"),
+			},
+		},
+		VolumeMounts: []corev1.VolumeMount{
+			{
+				Name:      naming.PVCTemplateName,
+				MountPath: naming.DataDir,
+				ReadOnly:  true,
+			},
+		},
+	}
 }
 
 func MakePodDisruptionBudget(sdc *scyllav1alpha1.ScyllaDBDatacenter) *policyv1.PodDisruptionBudget {